rename cosmo3D_v33.c to cosmo3D.c
